Added class to enable numpy overload, rewrote the code for it
Used a class to declare numbers then used custom math overloads so that we can use regular python typing within the equations. The logic of how python does math doesnt change in numpy the number handling is really what changes. All of the equations are rewritten with this in mind and when possible torque was updated. Currently the functions use global variables (filled with non-zero hunk), i would like to add inputs but without a solid understanding of the rest of the software structure i cannot. When using the Num class remember you have to *huh* dereference it with .out to get the final output.